+++ test/PHPStan/data/not-a-template.php
@@ -1,0 +1,7 @@
+<?php
+
+declare(strict_types=1);
+
+echo 'literal output';
+echo $someVariable;
+print $this;