+++ test/PHPStan/data/unescaped-exit.php
@@ -1,0 +1,23 @@
+<?php
+
+use SubstancePHP\HTTP\Renderer\HtmlRenderer;
+
+/**
+ * @var HtmlRenderer $this
+ * @var string $name
+ */
+if ($name === 'x') {
+    die('fatal error');
+}
+if ($name === 'x') {
+    exit;
+}
+if ($name === 'x') {
+    die($name);
+}
+if ($name === 'x') {
+    exit($this->raw($name));
+}
+if ($name === 'x') {
+    exit($this->h($name));
+}