+++ test/PHPStan/data/unescaped-print.php
@@ -1,0 +1,10 @@
+<?php
+
+use SubstancePHP\HTTP\Renderer\HtmlRenderer;
+
+/**
+ * @var HtmlRenderer $this
+ * @var string $name
+ */
+print $this->h($name);
+print $name;